Replaced DummyReadIndex with ReadIndexStub encapsulating the setup
